[AG-13] Nix environment, builds, and CI for frontend
The `shell.nix` file pins all dependencies in the environment to the versions
avilable in the top-level `pkgs.nix`, ensuring deterministic builds.

The Nix expression in `default.nix` provides a clean and optimized bundle in the
form of a Nix closure, for usage as part of higher level builds such as Docker
images.

The build process runs eslint and stylelint when building the closure, to
make sure nobody forgets to put it in a preCommit hook.

* Add missing dependencies for stylelint
* Provide build entry point through npm
* Write Nix derivation to build the closure with parcel