[AG-20] Add test which compares spec in the file and generated one
Problem: it's likely that somebody will update API and forget to update the
file spec manually, we need to prevent it.

Solution: add the test which generates a new Swagger API and compares it with
the one in the file. While doing this we had to change `Word`s for `Word32`s,
as the max bound for `Word` displays incorrectly after conversion to `Double`,
which happens before JSON encoding of the spec.