[AG-18] Use `loot-config` for configuration.
Problem: we need to somehow pass some configuration variables to our monad stack
Solution: use `loot-config` to define the config as a Vinyl record and put that
in the global reader data using the `Context` helper capability. Also, refactor
the web server code so that it uses the configuration variables correctly
while launching.