[AG-77] Post stub topic on Discourse
Problem: Agora backend has to be connected with Discourse
forum somehow.

Solution: as it was agreed (with Francis Brunelle)
Agora has to interact with Discourse in both directions:
post a topic about a proposal there as well as fetch info
about topics from there.
This commit introduces a client to interact with Discourse API,
posting topic stub on Discourse, when a new proposal is met in 'onApply',
also parsing of html and tests for it.