[AG-77] Move posting of Discourse topics to the separate worker
Problem: posting of a tpoic stub is performing
in the same thread where a block application happens.
It may cause a hang of syncing with the Tezos blockchain
if Discourse forum is anavailable, API token is outdated,
or an error happened due to Discourse changed its API, etc.

Solution: move a posting of topics to a separate worker
and push a new tasks for it when a new proposals are met.
This solution may cause a situation when a proposal was processed,
but due to errors posting on Discourse forum didn't happen, however,
I think the solution is a sane trade-off between the web site
stability and accomplishing the task of interaction with Discourse.