[AG-65] Handle fork in one block
Problem: we encounter a problem of divergent heads,
it happens because a block can't be treated valid until
a next one comes (which contains endorsements for the previous one).

Solution: block sync worker applies blocks
up to a previous of the passed head, not up
to the passed. So, the server will lag for one block
but I think it's acceptable.