[AG-141] Serve static files from a prefix
It is impossible to tell when a static file that we expect to have
(JS/CSS/etc) is missing because all requests for missing files are being
unconditionally redirected to index.html. However, sometimes we might
want to be able to tell the difference.

Solution: enforce a strict structure on the paths that we serve content
under, everything that is statically-generated is served under /static
and only there.

* Changes to the Caddy config:
  * Will rewrite /static/foo/bar -> /foo/bar and try to serve from the FS
  * If a static file does not exists on the FS, will return 404
  * Non-static files, as before, all point to index
* Changes to the frontend build process
  * Tell Parcel that files will be serverd from /static
  * The same CLI option should be passed to `serve`, but, unfortunately,
    it won’t work: https://github.com/parcel-bundler/parcel/issues/857
  * Looks like there is no way to add the point above as a comment to
    package.json because apparently JSON does not support coments
  * Looking at the progress on this issue above, I’d say we have to run
    away from this Parcel thing