[SDAO-49] Implement a QuorumGraph component
This component displays a horizontal bar that fills in as more voters
vote. It also draws a mark that shows the value of the quorum. The bar
is red before the quorum is reached and green after.

It is similar to the Supermajority graph in Tezos Agora’s promotion
period, except that it is conceptually simpler (and the implementation
is much more straightforward).