Allow to remove council members who voted
The votes table in the DB used to have a foreign key on (voter_pk,
stage) to the council table. Which means that if a council member has
voted in a given stage and then was removed, our council update function
would fail, as the foreign key constraint would become violated.

Thus, the constraint is removed.

Note: this way, the vote of the removed council member will still still
be displayed. To resolve this, when selecting valid votes for a given
stage, votes should be inner-joined on the council table.