[SDAO-46] Fix servant streaming regression
New streaming abstraction in servant was used incorrectly: the stream
(`SourceIO BlockHead`) was leaking outside the `withClientM` function
that cleans up the connection when the internal computation finishes.

The deep problem is that, in general, it is impossible to hoist a
streaming computation from `ClientM` to something like `IO`, precisely
because all the cleanup logic is contained withint the `ClientM` monad.

Solution:

1. Do not use `genericClientHoist` for `TezosClient`, as it contains
   streaming endpoints, which cannot be hoisted generically.
2. Instead manually convert standard `genericClient` to our
   `TezosClient` interface by hoisting simple endpoints as previously,
   but doing correct CPS on the streaming endpoint.