[SDAO-58] Never display more than 7 circles
The number of progress circles displayed was equal to the number of days
in the stage. This works as expected for stages that always have 7 days,
but the Implementation stage can have between 7 and 10, depending on the
month. The extra circles were causing the header to overflow.

Solution: never show more than 7 circles (in practice, always show
exactly 7). The circles are not rescaled, in that they still correspond
to days; days after the 7-th in an Implementation stage are invisible.