[SDAO-59] Add NixOS configuration module
This module is somewhat rudimentary in that it allows generating configs
using Nix and NixOS modules type system, however it does not actually
configure anything in NixOS.

The idea is to use it to generate a config that will be binded inside a
docker container. Later it can be extended to provide docker-ed and
docker-less NixOS service configurations.