Disable react-navi prefetch
Updating react-navi inroduced a regression: hovering mouse over stages
in the header caused them to load. The new version of rect-navi has some
link prefetch logic onMouseEnter, which, apparently, is supposed to load
the data used by the page to speed up the transtition. But in our case
it actually caused the view, probably, due to how the stage state is
treated.